it works v1

+++ micro-service-commande/src/main/java/com/example/micro_service_commande/SAGA/SagaMessage.java
@@ -39,7 +39,7 @@ public class SagaMessage implements Serializable {
         this.dispo = dispo;
     }
 
-    public long getPanierId() {
+    public int getPanierId() {
         return panierId;
     }
